+++ CHANGES.rst
@@ -1,11 +1,78 @@
-0.7.4 (unreleased)
+1.0.1 (unreleased)
 ==================
+
+Bug Fixes
+---------
+
+ramp_fitting
+~~~~~~~~~~~~
+
+- Updating multi-integration processing to correctly combine multiple
+  integration computations for the final image information. [#108]
+
+Changes to API
+--------------
+
+
+
+1.0.0 (2022-06-24)
+==================
+
+Bug Fixes
+---------
+
+ramp_fitting
+~~~~~~~~~~~~
+
+- Adding special case handler for GLS to handle one group ramps. [#97]
+
+- Updating how one group suppression and ZEROFRAME processing works with
+  multiprocessing, as well as fixing the multiprocessing failure. [#99]
+
+- Changing how ramp fitting handles fully saturated ramps. [#102]
+
+saturation
+~~~~~~~~~~
+
+- Modified the saturation threshold applied to pixels flagged with
+  NO_SAT_CHECK, so that they never get flagged as saturated. [#106]
+
+Changes to API
+--------------
+
+ramp_fitting
+~~~~~~~~~~~~
+
+- The tuple ``integ_info`` no longer returns ``int_times`` as a part of it,
+  so the tuple is one element shorter. [#99]
+
+- For fully saturated exposures, all returned values are ``None``, instead
+  of tuples. [#102]
+
+saturation
+~~~~~~~~~~~
+
+- Changing parameter name in twopoint_difference from 'normal_rej_thresh' to rejection_thresh' for consistency. [#105]
+
+Other
+-----
+
+general
+~~~~~~~
+
+- Update CI workflows to cache test environments and depend upon style and security checks [#96]
+- Increased required ``Python`` version from ``>=3.7`` to ``>=3.8`` (to align with ``astropy``) [#98]
 
 0.7.3 (2022-05-20)
 ==================
 
+jump
+~~~~
+
+- Update ``twopoint_difference.py`` [#90]
+
 ramp_fitting
-------------
+~~~~~~~~~~~~
 
 - Updating the one good group ramp suppression handler works. [#92]
 
@@ -13,7 +80,7 @@ ramp_fitting
 ==================
 
 ramp_fitting
-------------
+~~~~~~~~~~~~
 
 - Fix for accessing zero-frame in model to account for Roman data not using
   zero-frame. [#89]
@@ -23,25 +90,25 @@ ramp_fitting
 ==================
 
 jump
----- 
+~~~~
 - Enable multiprocessing for jump detection, which is controlled by the 'max_cores' parameter. [#87]
 
 0.7.0 (2022-05-13)
 ==================
 
 linearity
----------
+~~~~~~~~~
 - Added functionality to linearly process ZEROFRAME data the same way
   as the SCI data. [#81]
 
 ramp_fitting
-------------
+~~~~~~~~~~~~
 - Added functionality to use ZEROFRAME data in place of group 0 data
   for ramps that are fully saturated, but still have good ZEROFRAME
   data. [#81]
 
 saturation
----------- 
+~~~~~~~~~~
 - Added functionality to process ZEROFRAME data for saturation the same
   way as the SCI data. [#81]
 
@@ -50,7 +117,7 @@ saturation
 ==================
 
 saturation
----------- 
+~~~~~~~~~~
 
 - Added in functionality to deal with charge spilling from saturated pixels onto neighboring pixels [#83]
 
@@ -66,11 +133,11 @@ saturation
 ================
 
 jump
-----
+~~~~
 - Neighboring pixels with 'SATURATION' or 'DONOTUSE' flags are no longer flagged as jumps. [#79]
 
 ramp_fitting
-------------
+~~~~~~~~~~~~
 
 - Adding feature to use ZEROFRAME for ramps that are fully saturated, but
   the ZEROFRAME data for that ramp is good. [#81]
@@ -79,7 +146,7 @@ ramp_fitting
 ================
 
 ramp_fitting
-------------
+~~~~~~~~~~~~
 
 - Adding feature to suppress calculations for saturated ramps having only
   the 0th group be a good group.  [#76]
@@ -88,12 +155,12 @@ ramp_fitting
 ================
 
 ramp_fitting
-------------
+~~~~~~~~~~~~
 
 - Adding GLS code back to ramp fitting. [#64]
 
 jump
-----
+~~~~
 
 - Fix issue in jump detection that occured when there were only 2 usable
   differences with no other groups flagged. This PR also added tests and
@@ -103,7 +170,7 @@ jump
 ==================
 
 jump
-----
+~~~~
 
 - fixes to several existing errors in the jump detection step. added additional
   tests to ensure step is no longer flagging jumps for pixels with only two
@@ -113,7 +180,7 @@ jump
 ==================
 
 dark_current
-------------
+~~~~~~~~~~~~
 
 - Moved dark current code from JWST to STCAL. [#63]
 
@@ -121,22 +188,24 @@ dark_current
 ==================
 
 linearity
----------
+~~~~~~~~~
 - Let software set the pixel dq flag to NO_LIN_CORR if linear term of linearity coefficient is zero. [#65]
 
 ramp_fitting
-------------
+~~~~~~~~~~~~
 
 - Fix special handling for 2 group ramp. [#70]
 
 - Fix issue with inappropriately including a flagged group at the beginning
   of a ramp segment. [#68]
 
+- Changed Ramp Fitting Documentation [#61]
+
 0.4.2 (2021-10-28)
 ==================
 
 ramp_fitting
-------------
+~~~~~~~~~~~~
 
 - For slopes with negative median rates, the Poisson variance is zero. [#59]
 
@@ -147,7 +216,7 @@ ramp_fitting
 ==================
 
 jump_detection
---------------
+~~~~~~~~~~~~~~
 
 - Reverts "Fix issue with flagging for MIRI three and four group integrations. [#44]
 
@@ -156,21 +225,25 @@ jump_detection
 ==================
 
 jump_detection
---------------
+~~~~~~~~~~~~~~
 
 - Fix issue with flagging for MIRI three and four group integrations. [#44]
 
 linearity
----------
+~~~~~~~~~
 
 - Adds common code for linearity correction [#55]
 
+ramp_fitting
+~~~~~~~~~~~~
+
+- Global DQ variable removed [#54]
 
 0.3.0 (2021-09-28)
 ==================
 
 saturation
-----------
+~~~~~~~~~~
 
 - Adds common code for saturation [#39]
 
@@ -178,11 +251,15 @@ saturation
 0.2.5 (2021-08-27)
 ==================
 
+jump
+~~~~
+
+- added tests for two point difference [#37]
+
 ramp_fitting
-------------
+~~~~~~~~~~~~
 
-- Adds support for Roman ramp data. [#49]
-
+- Adds support for Roman ramp data. [#43] [#49]
 
 0.2.4 (2021-08-26)
 ==================
@@ -193,29 +270,36 @@ Workaround for setuptools_scm issues with recent versions of pip. [#45]
 0.2.3 (2021-08-06)
 ==================
 
-ramp_fitting
-------------
+jump
+~~~~
+- documentation changes + docs for jump detection [#14]
 
-- Fix ramp fitting multiprocessing. (#30)
+ramp_fitting
+~~~~~~~~~~~~
+
+- Fix ramp fitting multiprocessing. [#30]
 
 
 0.2.2 (2021-07-19)
 ==================
 
-ramp_fitting
-------------
-
-- Implemented multiprocessing for OLS. [#30]
-- Added DQ flag parameter to `ramp_fit` [#25]
+jump
+~~~~
 
 - Move common ``jump`` code to stcal [#27]
 
+ramp_fitting
+~~~~~~~~~~~~
+
+- Implemented multiprocessing for OLS. [#30]
+- Added DQ flag parameter to `ramp_fit` [#25]
+- Reduced data model dependency [#26]
 
 0.2.1 (2021-05-20)
 ==================
 
 ramp_fitting
-------------
+~~~~~~~~~~~~
 
 - Fixed bug for median ramp rate computation in report JP-1950. [#12]
 
@@ -224,7 +308,7 @@ ramp_fitting
 ==================
 
 ramp_fitting
-------------
+~~~~~~~~~~~~
 
 - Added ramp fitting code [#6]